adding so to use eject
