--- backend/main/api.py
@@ -1,19 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-import os
-
-basedir = os.path.abspath(os.path.dirname(__file__))
-
-# configuring the app
-app = Flask(__name__)
-app.config["SQLALCHEMY_DATABASE_URI"] = os.environ["DATABASE_URL"]
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-db = SQLAlchemy(app)
-
-# importing the app routes
-app.config.from_object(__name__)
-import views
-
-# if __name__ == '__main__':
-#     port = int(os.environ.get("PORT", 5000))
-#     app.run(debug=True, host="0.0.0.0", port=port)
