+++ docker-compose.yml
@@ -1,0 +1,16 @@
+version: '3.7'
+
+services:
+  backend:
+    build:
+      context: ./backend
+    ports:
+      - "5000:5000"
+  frontend:
+    build:
+      context: ./frontend
+    ports:
+      - 3000:3000
+    environment:
+      - CHOKIDAR_USEPOLLING=true  
+    stdin_open: true
