--- run_backend.py
@@ -1,5 +1,0 @@
-from backend import app
-import os
-
-port = int(os.environ.get("PORT", 5000))
-app.run(debug=True, host="0.0.0.0", port=port)
